pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcpfw"
version = "0.1.0"
description = "MCP Firewall — transparent policy enforcement proxy for MCP servers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Kaustubh Phatak"}]
keywords = ["mcp", "security", "firewall", "ai-agents", "proxy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Security",
    "Programming Language :: Python :: 3.10",
]
dependencies = ["pyyaml>=6.0"]

[project.scripts]
mcpfw = "mcpfw.cli:main"
